requirements.txt
fastapi
uvicorn[standard]
python-dotenv
redis
rq==1.15.1
openai>=1.16.0
